refactor: rename contact context file
